archivo movido y refactorizado(a consultar)
